do not use EXPECT_PACKAGE any more
As this tool is now general over the year, it has no reason to check
what the overarching package name is, or even that one exists. It
instead just trusts the config to provide the right implementation directory,
and ensures that _some_ `Cargo.toml` exists.

Closes #3.

Necessary updates to ensure that this compiled involved ensuring that
the CLI accepts the year, or defaults to the current year. That means
that we can close #4 also.